api/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2==2.9.6

uwsgi==2.0.21
django-cors-headers==4.2.0
